+++ mobile_frontend/lib/screens/authenticate/register.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:mobile_frontend/shared/field_widgets.dart';
 
 class Register extends StatefulWidget {
   @override
@@ -185,7 +186,6 @@ class _RegisterState extends State<Register> {
                         height: 50,
                         child: DropDownList(
                             comboList : <String>['Etudiant', 'Enseignant'],
-                            dropdownValue :'Etudiant'
                         ),
                       ),
                     ]
@@ -215,59 +215,6 @@ class _RegisterState extends State<Register> {
                   ),
                 ),
               ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class DropDownList extends StatefulWidget {
-  List<String> comboList;
-  String dropdownValue;
-
-  DropDownList({Key key, this.comboList, this.dropdownValue}) : super(key: key);
-
-  @override
-  _DropDownList createState() => _DropDownList();
-}
-
-class _DropDownList extends State<DropDownList> {
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 50,
-      child: Container(
-        decoration: BoxDecoration(
-          color: Colors.white,
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(10.0),
-          child: DropdownButtonHideUnderline (
-            child: DropdownButton<String>(
-              onTap: (){
-                FocusManager.instance.primaryFocus.unfocus();
-              },
-              value: widget.dropdownValue,
-              icon: Icon(Icons.keyboard_arrow_down),
-              iconSize: 40,
-              iconEnabledColor: Color(0xff43B485),
-              elevation: 16,
-              isExpanded: true,
-              style: TextStyle(color: Colors.black),
-              onChanged: (String newValue) {
-                setState(() {
-                  widget.dropdownValue = newValue;
-                });
-              },
-              items: widget.comboList.map<DropdownMenuItem<String>>((String value) {
-                return DropdownMenuItem<String>(
-                  value: value,
-                  child: Text(value),
-                );
-              }).toList(),
             ),
           ),
         ),
